refactor: 위치 이동
