+++ src/main/java/model/constant/LottoPrizeMoney.java
@@ -1,11 +1,12 @@
-package model;
+package model.constant;
 
 public enum LottoPrizeMoney {
     FIRST_PRIZE_MONEY(2 * 1000 * 1000 * 1000),
     SECOND_PRIZE_MONEY(30 * 1000 * 1000),
     THIRD_PRIZE_MONEY(1500 * 1000),
     FOURTH_PRIZE_MONEY(50 * 1000),
-    FIFTH_PRIZE_MONEY(5 * 1000);
+    FIFTH_PRIZE_MONEY(5 * 1000),
+    LOSE_PRIZE_MONEY(0);
     int value;
 
     LottoPrizeMoney(int value) {
